refactor(taxon): extract shared TaxonHeroCard component
TaxonDetail and TaxonDetailPanel rendered a byte-identical ~120-line
'image card + title block' — the 240x240 hero image, scientific/common
name, conservation + extinct chips, observation count, and GBIF/Wikidata
links. Extract a TaxonHeroCard component (with a Storybook story) and use
it in both, dropping the now-unused imports and gbifUrl/wikidataUrl locals.
