+++ frontend/src/components/taxon/TaxonDetail.tsx
@@ -8,7 +8,6 @@ import {
   Button,
   Stack,
   IconButton,
-  Chip,
   Divider,
   Link as MuiLink,
   List,
@@ -18,13 +17,12 @@ import {
   AccordionDetails,
 } from "@mui/material";
 import ArrowBackIcon from "@mui/icons-material/ArrowBack";
-import OpenInNewIcon from "@mui/icons-material/OpenInNew";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import { slugToName } from "../../lib/taxonSlug";
 import { useTaxon, useTaxonOccurrences } from "../../lib/query/hooks";
-import { ConservationStatus } from "../common/ConservationStatus";
 import { LoadMoreButton } from "../common/LoadMoreButton";
 import { TaxonLink, shouldItalicizeTaxonName } from "../common/TaxonLink";
+import { TaxonHeroCard } from "./TaxonHeroCard";
 import { usePageTitle } from "../../hooks/usePageTitle";
 import { useWikidataThumbnails } from "../../hooks/useWikidataThumbnails";
 import { WikiTaxonThumbnail } from "../common/WikiTaxonThumbnail";
@@ -111,10 +109,6 @@ export function TaxonDetail() {
     );
   }
 
-  // Use external URLs from the API response
-  const gbifUrl = taxon.gbifUrl;
-  const wikidataUrl = taxon.wikidataUrl;
-
   // Hero image: prefer photoUrl, fall back to larger Wikidata thumbnail
   const heroUrl =
     taxon.photoUrl ||
@@ -190,129 +184,7 @@ export function TaxonDetail() {
             </Stack>
           )}
 
-          {/* Image card + title block */}
-          <Stack
-            direction={{ xs: "column", sm: "row" }}
-            spacing={2}
-            sx={{ alignItems: { xs: "stretch", sm: "flex-start" } }}
-          >
-            {heroUrl && (
-              <Box
-                sx={{
-                  width: 240,
-                  height: 240,
-                  maxWidth: "100%",
-                  flexShrink: 0,
-                  borderRadius: 1,
-                  overflow: "hidden",
-                  backgroundColor: "action.hover",
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                  mx: { xs: "auto", sm: 0 },
-                }}
-              >
-                <Box
-                  component="img"
-                  src={heroUrl}
-                  alt={taxon.scientificName}
-                  sx={{
-                    width: "100%",
-                    height: "100%",
-                    objectFit: "contain",
-                    display: "block",
-                  }}
-                />
-              </Box>
-            )}
-            <Box sx={{ flex: 1, minWidth: 0 }}>
-              {/* Scientific Name */}
-              <Typography
-                variant="h5"
-                sx={{
-                  fontStyle: shouldItalicizeTaxonName(taxon.scientificName, taxon.rank)
-                    ? "italic"
-                    : "normal",
-                  color: "primary.main",
-                  fontWeight: 600,
-                }}
-              >
-                {taxon.scientificName}
-              </Typography>
-
-              {/* Common Name */}
-              {taxon.commonName && (
-                <Typography
-                  variant="h6"
-                  component="p"
-                  sx={{
-                    color: "text.secondary",
-                    mt: 0.5,
-                  }}
-                >
-                  {taxon.commonName}
-                </Typography>
-              )}
-
-              {/* Stats + External Links */}
-              <Stack
-                direction="row"
-                spacing={1}
-                sx={{
-                  alignItems: "center",
-                  mt: 2,
-                  flexWrap: "wrap",
-                }}
-              >
-                {taxon.conservationStatus && (
-                  <ConservationStatus status={taxon.conservationStatus} showLabel />
-                )}
-                {taxon.extinct && <Chip label="Extinct" size="small" color="error" />}
-                <Typography
-                  variant="body2"
-                  sx={{
-                    color: "text.secondary",
-                  }}
-                >
-                  {taxon.observationCount} observation{taxon.observationCount !== 1 ? "s" : ""} on
-                  Observ.ing
-                  {taxon.numDescendants !== undefined && taxon.numDescendants > 0 && (
-                    <> &middot; {taxon.numDescendants.toLocaleString()} descendant taxa</>
-                  )}
-                </Typography>
-                {(gbifUrl || wikidataUrl) && (
-                  <>
-                    {gbifUrl && (
-                      <Chip
-                        component="a"
-                        href={gbifUrl}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        label="GBIF"
-                        size="small"
-                        variant="outlined"
-                        clickable
-                        icon={<OpenInNewIcon sx={{ fontSize: 14 }} />}
-                      />
-                    )}
-                    {wikidataUrl && (
-                      <Chip
-                        component="a"
-                        href={wikidataUrl}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        label="Wikidata"
-                        size="small"
-                        variant="outlined"
-                        clickable
-                        icon={<OpenInNewIcon sx={{ fontSize: 14 }} />}
-                      />
-                    )}
-                  </>
-                )}
-              </Stack>
-            </Box>
-          </Stack>
+          <TaxonHeroCard taxon={taxon} heroUrl={heroUrl} />
 
           {/* Taxonomy Tree */}
           {((taxon.ancestors?.length ?? 0) > 0 || (taxon.children?.length ?? 0) > 0) && (
